PAYARA-4083: Simplify YubicoAPI to be a simple class
